Refactor cli.js to main.js
