+++ package.json
@@ -26,7 +26,7 @@
     "node": ">=4"
   },
   "scripts": {
-    "test": "xo && nyc ava",
+    "test": "./main.js && xo && nyc ava",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "files": [
